detector test utils (new module)
