+++ aderyn_core/src/detect/high/block_timestamp_deadline.rs
@@ -115,14 +115,11 @@ impl IssueDetector for BlockTimestampDeadlineDetector {
 
 #[cfg(test)]
 mod block_timestamp_deadline_detector_tests {
-    use crate::detect::{
-        detector::{detector_test_helpers::load_contract, IssueDetector},
-        high::BlockTimestampDeadlineDetector,
-    };
+    use crate::detect::{detector::IssueDetector, high::BlockTimestampDeadlineDetector};
 
     #[test]
     fn test_block_timestamp_deadline_uniswap_v2_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/UniswapV2Swapper.sol/UniswapV2Swapper.json",
         );
 
@@ -154,7 +151,7 @@ mod block_timestamp_deadline_detector_tests {
 
     #[test]
     fn test_block_timestamp_deadline_uniswap_v3_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/UniswapV3Swapper.sol/UniswapV3Swapper.json",
         );
