+++ aderyn_core/src/detect/high/delegate_call_in_loop.rs
@@ -66,13 +66,13 @@ impl IssueDetector for DelegateCallInLoopDetector {
 
 #[cfg(test)]
 mod delegate_call_in_loop_detector_tests {
-    use crate::detect::detector::{detector_test_helpers::load_contract, IssueDetector};
+    use crate::detect::detector::IssueDetector;
 
     use super::DelegateCallInLoopDetector;
 
     #[test]
     fn test_delegate_call_in_loop_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/ExtendedInheritance.sol/ExtendedInheritance.json",
         );
