+++ aderyn_core/src/detect/low/_template.rs
@@ -53,14 +53,11 @@ impl IssueDetector for TemplateDetector {
 
 #[cfg(test)]
 mod template_detector_tests {
-    use crate::detect::{
-        detector::{detector_test_helpers::load_contract, IssueDetector},
-        low::template_detector::TemplateDetector,
-    };
+    use crate::detect::{detector::IssueDetector, low::template_detector::TemplateDetector};
 
     #[test]
     fn test_template_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/ArbitraryTransferFrom.sol/ArbitraryTransferFrom.json",
         );
