+++ aderyn_core/src/detect/low/deprecated_oz_functions.rs
@@ -83,13 +83,13 @@ impl IssueDetector for DeprecatedOZFunctionsDetector {
 
 #[cfg(test)]
 mod deprecated_oz_functions_tests {
-    use crate::detect::detector::{detector_test_helpers::load_contract, IssueDetector};
+    use crate::detect::detector::IssueDetector;
 
     use super::DeprecatedOZFunctionsDetector;
 
     #[test]
     fn test_deprecated_oz_functions_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/DeprecatedOZFunctions.sol/DeprecatedOZFunctions.json",
         );
