+++ aderyn_core/src/detect/low/non_reentrant_before_others.rs
@@ -52,15 +52,13 @@ impl IssueDetector for NonReentrantBeforeOthersDetector {
 
 #[cfg(test)]
 mod non_reentrant_before_others_tests {
-    use crate::detect::{
-        detector::{detector_test_helpers::load_contract, IssueDetector},
-        low::NonReentrantBeforeOthersDetector,
-    };
+    use crate::detect::{detector::IssueDetector, low::NonReentrantBeforeOthersDetector};
 
     #[test]
     fn test_non_reentrant_before_others() {
-        let context =
-            load_contract("../tests/contract-playground/out/AdminContract.sol/AdminContract.json");
+        let context = crate::detect::test_utils::load_contract(
+            "../tests/contract-playground/out/AdminContract.sol/AdminContract.json",
+        );
 
         let mut detector = NonReentrantBeforeOthersDetector::default();
         let found = detector.detect(&context).unwrap();
