+++ aderyn_core/src/detect/low/push_0_opcode.rs
@@ -101,11 +101,11 @@ impl IssueDetector for PushZeroOpcodeDetector {
 
 #[cfg(test)]
 mod unspecific_solidity_pragma_tests {
-    use crate::detect::detector::{detector_test_helpers::load_contract, IssueDetector};
+    use crate::detect::detector::IssueDetector;
 
     #[test]
     fn test_push_0_opcode_detector_on_0_8_20() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/ExtendedInheritance.sol/ExtendedInheritance.json",
         );
 
@@ -136,8 +136,9 @@ mod unspecific_solidity_pragma_tests {
 
     #[test]
     fn test_push_0_opcode_detector_on_range() {
-        let context =
-            load_contract("../tests/contract-playground/out/CrazyPragma.sol/CrazyPragma.json");
+        let context = crate::detect::test_utils::load_contract(
+            "../tests/contract-playground/out/CrazyPragma.sol/CrazyPragma.json",
+        );
 
         let mut detector = super::PushZeroOpcodeDetector::default();
         let found = detector.detect(&context).unwrap();
@@ -166,7 +167,7 @@ mod unspecific_solidity_pragma_tests {
 
     #[test]
     fn test_push_0_opcode_detector_on_0_8_19() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/ArbitraryTransferFrom.sol/ArbitraryTransferFrom.json",
         );
 
@@ -180,8 +181,9 @@ mod unspecific_solidity_pragma_tests {
 
     #[test]
     fn test_push_0_opcode_detector_on_caret_0_8_13() {
-        let context =
-            load_contract("../tests/contract-playground/out/Counter.sol/Counter.0.8.25.json");
+        let context = crate::detect::test_utils::load_contract(
+            "../tests/contract-playground/out/Counter.sol/Counter.0.8.25.json",
+        );
 
         let mut detector = super::PushZeroOpcodeDetector::default();
         let found = detector.detect(&context).unwrap();
@@ -193,7 +195,7 @@ mod unspecific_solidity_pragma_tests {
 
     #[test]
     fn test_push_0_opcode_detector_on_greter_equal_0_8_0() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/IContractInheritance.sol/IContractInheritance.0.8.25.json",
         );
