+++ aderyn_core/src/detect/low/require_with_string.rs
@@ -56,13 +56,13 @@ impl IssueDetector for RequireWithStringDetector {
 
 #[cfg(test)]
 mod require_with_string_tests {
-    use crate::detect::detector::{detector_test_helpers::load_contract, IssueDetector};
+    use crate::detect::detector::IssueDetector;
 
     use super::RequireWithStringDetector;
 
     #[test]
     fn test_require_with_string() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/DeprecatedOZFunctions.sol/DeprecatedOZFunctions.json",
         );
