+++ aderyn_core/src/detect/low/useless_modifier.rs
@@ -56,13 +56,13 @@ impl IssueDetector for UselessModifierDetector {
 
 #[cfg(test)]
 mod useless_modifier_tests {
-    use crate::detect::detector::{detector_test_helpers::load_contract, IssueDetector};
+    use crate::detect::detector::IssueDetector;
 
     use super::UselessModifierDetector;
 
     #[test]
     fn test_useless_modifier_tests() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/OnceModifierExample.sol/OnceModifierExample.json",
         );
