+++ aderyn_core/src/detect/low/zero_address_check.rs
@@ -188,15 +188,12 @@ mod zero_address_check_tests {
     use crate::{
         ast::NodeType,
         context::{browser::GetClosestAncestorOfTypeX, workspace_context::ASTNode},
-        detect::{
-            detector::{detector_test_helpers::load_contract, IssueDetector},
-            low::ZeroAddressCheckDetector,
-        },
+        detect::{detector::IssueDetector, low::ZeroAddressCheckDetector},
     };
 
     #[test]
     fn test_deprecated_oz_functions_detector() {
-        let context = load_contract(
+        let context = crate::detect::test_utils::load_contract(
             "../tests/contract-playground/out/ZeroAddressCheck.sol/ZeroAddressCheck.json",
         );
