+++ aderyn_core/src/detect/mod.rs
@@ -4,6 +4,9 @@ pub mod helpers;
 pub mod high;
 pub mod low;
 
+#[cfg(test)]
+pub mod test_utils;
+
 #[macro_export]
 macro_rules! capture {
     ($self:ident, $context:ident, $item:expr) => {
